refactor(task): extract stage expansion and rerouting into graphresolver_stage.go
